third_party/WebKit/LayoutTests/accessibility/color-well-expected.txt

CONSOLE WARNING: line 10: The specified value "purple" does not conform to the required format.  The format is "#rrggbb" where rr, gg, bb are two-digit hexadecimal numbers.
This test ensures that the correct accessibility properties can be retrieved from input elements of type color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axColorwell.role is "AXRole: AXColorWell"
PASS axColorwell.colorValue is 4278190080
PASS axColorwell.stringValue is "AXValue: rgba(0, 0, 0, 1)"
PASS axColorwell.role is "AXRole: AXColorWell"
PASS axColorwell.colorValue is 4294901760
PASS axColorwell.stringValue is "AXValue: rgba(255, 0, 0, 1)"
PASS axColorwell.role is "AXRole: AXColorWell"
PASS axColorwell.colorValue is 4278190080
PASS axColorwell.stringValue is "AXValue: rgba(0, 0, 0, 1)"
PASS successfullyParsed is true

TEST COMPLETE

  
